pyproject.toml
[tool.poetry]
package-mode = false
name = "pyconmy24"
version = "0.1.0"
description = ""
authors = ["Nikho Sagala <nikho.sagala@jublia.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langchain = "^0.1.16"
langchain-openai = "^0.1.3"
langserve = {extras = ["all"], version = "^0.1.0"}
python-dotenv = "^1.0.1"
pydantic = "1.10.13"
requests = "^2.32.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
